+++ ChangeLog.md
@@ -10,6 +10,7 @@ Eventum Issue Tracking System
 - Fix 'get email' command in CLI (@balsdorf)
 - Fix bug with page specific JS not being called (@balsdorf)
 - Fix bug where session variable was never returned (@balsdorf)
+- Add project selection to CLI getWeeklyReport (@balsdorf)
 
 2016-06-06, Version [3.1.2]
 ----------------------------